Moved sh_*.c shaders from rt/ to liboptical/ directory. Moved and split rt/material.h into h/shadefuncs.h and h/shadework.h Changed calling sequence of mlib_setup() to get shader list head explicitly. Initialization of shader library is now via optical_shader_init().
Note that the "proper" location of shade.c, material.c, and refract.c
(presently in the rt/ directory) are a bit muddled, as they will be
shared with libmultispectral/ shader functions, but aren't quite
wholesome and pure yet.  To minimize impact to Lee's ongoing shader
developments, they haven't been changed.


git-svn-id: svn://svn.code.sf.net/p/brlcad/code/brlcad/trunk@15594 2f96ce8b-6d43-0410-b8df-bffccc660ffb