Fixed the 2 bugs that Lee detected while doing the Impact Flash animation:
1)  Progressively fewer instanced solids were having their stp
re-referenced because d_uses was not being zeroed by the "clean" command
in the animation script at the start of each new frame.

2) rt_clean() wasn't calling rt_clean_resource() for the per-CPU
resource structures (including the list of bu_bitv structures used for
re_solid_bitv) because of using the wrong macro to see if there were any
to do.  While fixing it I changed tactics so that a semaphored critical
section is no longer needed for this.


git-svn-id: svn://svn.code.sf.net/p/brlcad/code/brlcad/trunk@17662 2f96ce8b-6d43-0410-b8df-bffccc660ffb