Made significant updates to the human.c file which will eventually be a complete human generator that uses CSG to model a human.
human.c currently creates a 'stick' human to a desired height by use of command line parameters and also
allows for the output of specifically named .g file when using the -o command on input.
Also included are most of the bounding blocked for each part of the generated body by using the -b command.
So if you wanted to make a 'human' right now, the command would be...
./human -H60 -o man.g -b
And that would generate a 60 inch tall stick figure named man.g with bounding boxes.
the -? and -h commands bring up the help if needed

As for the code itself, each part of the person is divded up into its own function which creates that
part. So there is a makeHead which makes the head, a makeLeftFoot which makes the left foot and so on.
All funtions are controlled by the makeBody funtion which holds all relevant values to be sent to each
part funtion, and makeBody is called by main and is sent the values for the standing height of the person,
and whether there should be bounding boxes included.



git-svn-id: svn://svn.code.sf.net/p/brlcad/code/brlcad/trunk@34849 2f96ce8b-6d43-0410-b8df-bffccc660ffb